+++ CHANGES.rst
@@ -14,6 +14,8 @@ https://github.com/zopefoundation/Zope/blob/4.0a6/CHANGES.rst
 Bugfixes
 ++++++++
 
+- Install a default page for the root view in new installations again.
+
 - Re-raise app exceptions if wsgi.handleErrors is False in the request environ.
 
 - Fix path expressions trying to call views that do not implement `__call__`.